switch to using a wizard
